.history/requirements_20251113234825.txt
blinker==1.9.0
click==8.3.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
pytest
pytest-cov

# FastAPI et dépendances
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Authentification
PyJWT==2.8.0
passlib==1.7.4
python-jose==3.3.0

# Files d'attente et traitement asynchrone
celery==5.3.4
redis==5.0.1

# Documentation OpenAPI
fastapi-users==12.1.4
fastapi-versioning==0.1.2
